+++ nava/functions.py
@@ -4,8 +4,11 @@ import sys
 import subprocess
 import os
 import shlex
+from functools import wraps
 
-from .params import OVERVIEW, SOUND_FILE_PLAY_ERROR, SOUND_FILE_EXIST_ERROR, SOUND_FILE_PATH_TYPE_ERROR
+from .params import OVERVIEW
+from .params import SOUND_FILE_PLAY_ERROR, SOUND_FILE_EXIST_ERROR
+from .params import SOUND_FILE_PATH_TYPE_ERROR
 from .errors import NavaBaseError
 
 
@@ -26,21 +29,22 @@ def quote(func):
 
     :return: inner function
     """
-    def inner_function(*args, **kwargs):
+    @wraps(func)
+    def quoter(sound_path, *args, **kwargs):
         """
         Inner function.
 
+        :param sound_path: sound path
+        :type sound_path: str
         :param args: non-keyword arguments
         :type args: list
         :param kwargs: keyword arguments
         :type kwargs: dict
         :return: modified function result
         """
-        sound_path = args[0]
         sound_path = shlex.quote(sound_path)
-        args[0] = sound_path
-        return func(*args, **kwargs)
-    return inner_function
+        return func(sound_path, *args, **kwargs)
+    return quoter
 
 
 def __play_win(sound_path):
@@ -95,24 +99,26 @@ def path_check(func):
 
     :return: inner function
     """
-    def inner_function(*args, **kwargs):
+    @wraps(func)
+    def path_checker(sound_path, *args, **kwargs):
         """
         Inner function.
 
+        :param sound_path: sound path
+        :type sound_path: str
         :param args: non-keyword arguments
         :type args: list
         :param kwargs: keyword arguments
         :type kwargs: dict
         :return: modified function result
         """
-        sound_path = args[0]
-        if not (isinstance(sound_path, str)):
+        if not isinstance(sound_path, str):
             raise NavaBaseError(SOUND_FILE_PATH_TYPE_ERROR)
         # check sound file existance
-        if not (os.path.isfile(sound_path)):
+        if not os.path.isfile(sound_path):
             raise NavaBaseError(SOUND_FILE_EXIST_ERROR)
-        return func(*args, **kwargs)
-    return inner_function
+        return func(sound_path, *args, **kwargs)
+    return path_checker
 
 
 @path_check